requirements.txt
bcrypt==1.1.1
cffi==0.9.2
Django==1.8
django-debug-toolbar==1.3.0
django-filter==0.9.2
django-suit==0.2.13
mysqlclient==1.3.6
Pillow==2.8.1
pycparser==2.12
six==1.9.0
sqlparse==0.1.15
